API: Upgrade to phpstan 0.12
Unfortunately retaining 0.11 support isn’t feasible but that seems like
an acceptable tradeoff.

Note that some of the type-hinting of DataLists will now be doable with
phpstan’s template features - we can type hint DataList<Member> for
example, and we can commit some @phpstan-return clauses to the core
framework to support this.